Get_IP_Address_Hostname
IP Address  Stands for Internet Protocol Address, is a numeric label that is assigned uniquely to each device connected within a computer network that uses the internet protocol. An IP address serves two principal functions:

It identifies the host, or more specifically its network interface.
It provides the location of the host in the network, and thus the capability of establishing a path to that host.